+++ src/components/UserPage/Video/index.tsx
@@ -2,28 +2,11 @@ import { useState } from 'react';
 import * as S from './style';
 import * as I from 'assets/svg';
 
-function UserPage() {
+function VideoSection() {
   const [activeIndex, setActiveIndex] = useState<number>(0);
 
   return (
-    <S.UserPageLayout>
-      <S.UserInfoSection>
-        <S.UserProfile src="/images/background.png" />
-        <S.UserValues>
-          <S.UserName>오종진입니다</S.UserName>
-          <S.Follow>
-            <S.FollowBox>
-              <p className="key">팔로잉</p>
-              <S.Value>102</S.Value>
-            </S.FollowBox>
-            <S.FollowBox>
-              <p className="key">팔로워</p>
-              <S.Value>102</S.Value>
-            </S.FollowBox>
-          </S.Follow>
-        </S.UserValues>
-      </S.UserInfoSection>
-      <S.FollowButton>팔로우</S.FollowButton>
+    <>
       <S.Category>
         <S.Box
           onClick={() => setActiveIndex(0)}
@@ -48,8 +31,8 @@ function UserPage() {
         </S.Box>
       </S.Category>
       <S.VideoSection />
-    </S.UserPageLayout>
+    </>
   );
 }
 
-export default UserPage;
+export default VideoSection;
