+++ src/pages/User.tsx
@@ -1,5 +1,5 @@
 import Header from 'components/Common/Header';
-import UserPage from 'components/UserPage';
+import UserPage from 'components/UserPage/Page';
 
 function User() {
   return (
